Moved ens validation to util
